==> Preterm Bayesian Modelling Code/0.fileNames.R <==
wpp2<-"WPP2022_Live_births_by_sex.xlsx"
regionsName<- "sbr_regions v2.csv"
birth_healthfacility<- "birth_healthfacility_who2021.dta"

adminName<-"inputs/Admin_Database_04March2022.xlsx"

surveyPNameNew<-"inputs/pretermStudies_16112022_hb2_14thdec.xlsx"
